% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probe2symbol_df}
\alias{probe2symbol_df}
\title{a data.frame for converting probe ids to gene symbols}
\format{a data.frame with 23749 rows and 2 column}
\usage{
probe2symbol_df
}
\description{
the first column is a vector of probe ids
}
\details{
the other columns is a vector of gene symbols
}
\keyword{datasets}
